brand.ts and messages.ts
